qr_code_service/requirements.txt
falcon==1.4.1
google-cloud-storage==1.13.1
requests==2.21.0
requests-toolbelt==0.8.0
qrcode==6.1
